# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Drive Backup Console — Agent API",
    "version": "1.0.0",
    "description": "Programmatic file API for AI agents and scripts. Authenticate with an API key via `Authorization: Bearer <token>` (or a browser session cookie for the UI). Read-only keys may call `read` operations; `readwrite` keys may also mutate. Keys act as the site owner against the server-side Google Drive token; scope limits what a key may do, not whose Drive it sees."
  },
  "servers": [{ "url": "/api/v1" }],
  "security": [{ "BearerAuth": [] }, { "CookieAuth": [] }],
...
  "paths": {
...
    "/files": {
      "get": {
        "tags": ["files"], "summary": "List folder children", "operationId": "listFiles",
        "x-required-scope": "read",
        "parameters": [
          { "name": "folderId", "in": "query", "schema": { "type": "string" }, "description": "Folder id; defaults to the configured root" },
          { "name": "pageToken", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Page of items", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListResult" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      },
      "post": {
        "tags": ["files"], "summary": "Create a file with text content", "operationId": "createFile",
        "x-required-scope": "readwrite",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "parentId": { "type": "string" }, "content": { "type": "string" }, "mimeType": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Created file", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FileItem" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
...
    "/files/{id}": {
      "patch": {
        "tags": ["files"], "summary": "Update name and/or description", "operationId": "updateFile",
        "x-required-scope": "readwrite",
        "description": "At least one of name/description is required. An empty description clears it. Folder descriptions help agents understand a folder's purpose and are returned by list/search.",
        "parameters": [{ "$ref": "#/components/parameters/PathId" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "description": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Updated item", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FileItem" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "delete": {
        "tags": ["files"], "summary": "Move a file to trash", "operationId": "trashFile",
        "x-required-scope": "readwrite",
        "parameters": [{ "$ref": "#/components/parameters/PathId" }],
        "responses": { "204": { "description": "Trashed" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/files/{id}/content": {
      "get": {
        "tags": ["files"], "summary": "Read text file content", "operationId": "getContent",
        "x-required-scope": "read",
        "parameters": [{ "$ref": "#/components/parameters/PathId" }],
        "responses": { "200": { "description": "File content" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "put": {
        "tags": ["files"], "summary": "Overwrite text file content", "operationId": "putContent",
        "x-required-scope": "readwrite",
        "parameters": [{ "$ref": "#/components/parameters/PathId" }],
        "requestBody": { "required": true, "content": { "text/plain": { "schema": { "type": "string" } } } },
        "responses": { "200": { "description": "Updated" }, "403": { "$ref": "#/components/responses/Forbidden" }, "413": { "$ref": "#/components/responses/TooLarge" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "description": "API key token: `Authorization: Bearer dbc_...`" },
      "CookieAuth": { "type": "apiKey", "in": "cookie", "name": "dbc_session", "description": "Browser session cookie (UI). Full read-write." }
    },
    "parameters": {
      "PathId": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Drive file/folder id" }
    },
    "schemas": {
...
      "Error": { "type": "object", "properties": { "error": { "type": "object", "properties": { "code": { "type": "string" }, "message": { "type": "string" } } } } },
      "FileItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "mimeType": { "type": "string" },
          "size": { "type": ["integer", "null"] },
          "modifiedTime": { "type": "string" },
          "isFolder": { "type": "boolean" },
          "description": { "type": "string", "description": "Drive native description (folder/file blurb)" },
          "thumbnailUrl": { "type": "string" }
        }
      },
      "ListResult": {
        "type": "object",
        "properties": {
          "folderId": { "type": "string" },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/FileItem" } },
          "nextPageToken": { "type": "string" }
        }
      },
...
    },
    "responses": {
      "BadRequest": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unauthorized": { "description": "Missing or invalid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Forbidden": { "description": "API key lacks the required scope", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "Resource not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "TooLarge": { "description": "Request body exceeds the size limit", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    }
  }
}